Added Squareword display & CLI
