+++ Wordle.Core/SquareWord.cs
@@ -7,16 +7,25 @@ public class SquareWord : BaseGame
 
     public char[,] Grid { get; }
     public IReadOnlyList<string> SelectedWords { get; }
-    public SquareWord(int wordLength, int possibleTries, WordList list)
+    public char[,] WellPlacedLetters { get; }
+    public ImmutableHashSet<char>[] WronglyPlacedLetters { get; }
+    public ImmutableHashSet<char> WrongLetters { get; private set; }
+    public SquareWord(int wordLength, int possibleTries, WordList list, int? seed)
         : base(wordLength, possibleTries, false, list)
     {
-        TryAddVertical(0, ImmutableList.Create<string>(), ImmutableList.Create<string>(), out var selectedHorizontalWords, new());
+        TryAddVertical(0, ImmutableList.Create<string>(), ImmutableList.Create<string>(), out var selectedHorizontalWords, seed is int i ? new(i) : new());
 
         SelectedWords = selectedHorizontalWords.ToList().AsReadOnly();
+        WellPlacedLetters = new char[WordLength, WordLength];
+        WronglyPlacedLetters = new ImmutableHashSet<char>[WordLength];
+        WrongLetters = ImmutableHashSet<char>.Empty;
         Grid = new char[WordLength, WordLength];
         for (int y = 0; y < WordLength; y++)
+        {
+            WronglyPlacedLetters[y] = ImmutableHashSet<char>.Empty;
             for (int x = 0; x < WordLength; x++)
                 Grid[x, y] = SelectedWords[y][x];
+        }
 
         bool TryAddHorizontal(int i, ImmutableList<string> selectedVerticalWords, ImmutableList<string> selectedHorizontalWords, out ImmutableList<string> horizontalWords, Random rng, bool dryRun = false)
         {
@@ -63,6 +72,9 @@ public class SquareWord : BaseGame
         }
     }
 
+    public override SquareWord Recreate()
+        => new(WordLength, PossibleTries, CompleteWordList, null);
+
     private (int offset, int length)? GetWordListOffsetAndLength(string start)
     {
         var range = GetBound<string>(WordList.AsArray(), w => w.AsSpan().StartsWith(start));
@@ -91,5 +103,37 @@ public class SquareWord : BaseGame
         if (negate)
             return null;
         return start..end;
+    }
+
+    public bool Try(string word)
+    {
+        if (RemainingTries > 0 || word.Length != WordLength || !WordList.Contains(word))
+            return false;
+        foreach (var letter in word)
+        {
+            if (!SelectedWords.SelectMany(static w => w).Contains(letter))
+                WrongLetters = WrongLetters.Add(letter);
+        }
+        for (int y = 0; y < WordLength; y++)
+        {
+            var letterCounts = word.GroupBy(static c => c)
+                .ToDictionary(static g => g.Key, g => word.Count(c => c == g.Key));
+            for (int x = 0; x < WordLength; x++)
+                if (!WrongLetters.Contains(word[x]) && SelectedWords[y][x] == word[x] && letterCounts.TryGetValue(word[x], out var count) && count > 0)
+                {
+                    WronglyPlacedLetters[y] = WronglyPlacedLetters[y].Remove(word[x]);
+                    letterCounts[word[x]]--;
+                    WellPlacedLetters[x, y] = word[x];
+                }
+            for (int x = 0; x < WordLength; x++)
+                if (!WrongLetters.Contains(word[x]) && SelectedWords[y].Contains(word[x]) && letterCounts.TryGetValue(word[x], out var count) && count > 0)
+                {
+                    letterCounts[word[x]]--;
+                    WronglyPlacedLetters[y] = WronglyPlacedLetters[y].Add(word[x]);
+                }
+        }
+        RemainingTries--;
+
+        return true;
     }
 }
